refactor(billing): extract stripe rollback helpers
